+++ src/onegov.form/onegov/form/collection.py
@@ -1,8 +1,14 @@
 from delorean import Delorean
 from datetime import datetime, timedelta
 from onegov.core.utils import normalize_for_url
-from onegov.form.models import FormDefinition, FormSubmission
-from sqlalchemy import inspect
+from onegov.form.fields import UploadField
+from onegov.form.models import (
+    FormDefinition,
+    FormSubmission,
+    FormSubmissionFile
+)
+from sqlalchemy import inspect, not_
+from uuid import uuid4
 
 
 class FormCollection(object):
@@ -94,19 +100,13 @@ class FormSubmissionCollection(object):
 
         # look up the right class depending on the type
         _mapper = inspect(FormSubmission).polymorphic_map.get(state)
+
         submission = (_mapper and _mapper.class_ or FormSubmission)()
+        submission.id = uuid4()
         submission.name = form_name
-        submission.definition = form._source
-        submission.data = form.data
         submission.state = state
 
-        # pending submissions are not necessarily valid, however we don't need
-        # to store invalid state as it is wiped out anyway
-        submission.prune(form)
-
-        # never include the csrf token
-        if form.meta.csrf and form.meta.csrf_field_name in submission.data:
-            del submission.data[form.meta.csrf_field_name]
+        self.update(submission, form)
 
         self.session.add(submission)
         self.session.flush()
@@ -119,6 +119,71 @@ class FormSubmissionCollection(object):
         )
 
         return submission
+
+    def update(self, submission, form):
+        """ Takes a submission and a form and updates the submission data
+        as well as the files stored in a spearate table.
+
+        """
+        assert submission.id and submission.state
+
+        submission.definition = form._source
+        submission.data = form.data
+
+        # pending submissions are not necessarily valid, however we don't need
+        # to store invalid state as it is wiped out anyway
+        submission.prune(form)
+
+        # never include the csrf token
+        if form.meta.csrf and form.meta.csrf_field_name in submission.data:
+            del submission.data[form.meta.csrf_field_name]
+
+        # move uploaded files to a separate table
+        files = set(
+            field_id for field_id, field in form._fields.items()
+            if isinstance(field, UploadField)
+        )
+
+        files_to_remove = set(
+            id for id in files
+            if submission.data.get(id) == {}
+        )
+
+        files_to_add = set(
+            id for id in (files - files_to_remove)
+            if submission.data.get(id)
+            and not submission.data[id]['data'].startswith('@')
+        )
+
+        files_to_keep = files - files_to_remove - files_to_add
+
+        # delete all files which are not part of the updated form
+        # if no files are given, delete all files belonging to the submission
+        query = self.session.query(FormSubmissionFile)
+        query = query.filter(FormSubmissionFile.submission_id == submission.id)
+
+        if files_to_keep:
+            query = query.filter(not_(
+                FormSubmissionFile.field_id.in_(files_to_keep)))
+
+        query.delete('fetch')
+
+        # store the new fields in the separate table
+        for field_id in files_to_add:
+            f = FormSubmissionFile(
+                id=uuid4(),
+                field_id=field_id,
+                submission_id=submission.id,
+                filedata=submission.data[field_id]['data'],
+            )
+            self.session.add(f)
+
+            # replace the data in the submission with a reference
+            submission.data[field_id]['data'] = '@{}'.format(f.id.hex)
+
+            # we need to mark these changes as only top-level json changes
+            # are automatically propagated
+            submission.data.changed()
 
     def remove_old_pending_submissions(self, older_than):
         """ Removes all pending submissions older than the given date. The